Seperate DB operations from Network.swift to DBManager.swift
